// frontend/src/types/tmdb.ts
export type TrendingMedia = {
  id: number;
  title?: string;
  name?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  media_type?: "movie" | "tv" | "person";
  vote_average?: number;
  overview?: string;
  release_date?: string;
  first_air_date?: string;
  genre_ids?: number[];
  original_language?: string;
};